chore: restructured components, made folders names more readable
